add validations before deleting the ephemeral table

+++ core/src/main/scala/org/apache/spark/sql/crossdata/catalog/ZookeeperStreamingCatalog.scala
@@ -49,6 +49,12 @@ class ZookeeperStreamingCatalog(xdContext: XDContext) extends XDStreamingCatalog
     ephemeralTableDAO.dao.update(ephemeralTable.name, ephemeralTable)
 
   override def dropEphemeralTable(tableIdentifier: String): Unit = {
+    val isRunning = ephemeralTableStatusDAO.dao.get(tableIdentifier).map{ tableStatus =>
+      tableStatus.status == EphemeralExecutionStatus.Started || tableStatus.status == EphemeralExecutionStatus.Starting
+    } getOrElse notFound(tableIdentifier)
+
+    if(isRunning) throw new RuntimeException("The ephemeral is running. The process should be stopped first using 'Stop <tableIdentifier>'")
+    
     ephemeralTableDAO.dao.delete(tableIdentifier)
     ephemeralTableStatusDAO.dao.delete(tableIdentifier)
   }
@@ -109,5 +115,7 @@ class ZookeeperStreamingCatalog(xdContext: XDContext) extends XDStreamingCatalog
 
   override protected[crossdata] def dropAllEphemeralStatus(): Unit =
     ephemeralTableStatusDAO.dao.deleteAll
+
+
 
 }
